WIP: shimv2 tracing
Added tracing with tracer being created in Start, which should ideally
be moved to New or Create. Currently getting no trace or span output
when the tracer is created in New, Create, or StartShim, or any log
output from those functions even though they are being called (as seen
by writing to file).

To do:
 - Test with ctx saved in service struct (which is the same for the
duration of a container) vs ctx passed in (which is different, at least
by memory address, each time).
 - Test with saved ctx starting in Create and then New
 - Enable containerd debugging
 - Possibly update containerd